+++ components/admin_console/admin_definition.jsx
@@ -31,7 +31,7 @@ import PermissionTeamSchemeSettings from './permission_schemes_settings/permissi
 import SystemUsers from './system_users';
 import ServerLogs from './server_logs';
 import BrandImageSetting from './brand_image_setting/brand_image_setting.jsx';
-import GroupSettings from './group_settings';
+import GroupSettings from './group_settings/group_settings.jsx';
 
 import * as DefinitionConstants from './admin_definition_constants';
