+++ package.json
@@ -28,7 +28,7 @@
     "localforage": "1.7.2",
     "localforage-observable": "1.4.0",
     "marked": "github:mattermost/marked#ed33baecd7d7fa97d479ba22dde9d226b083d67d",
-    "mattermost-redux": "github:mattermost/mattermost-redux#9ce97c0b659024a1557e5c90dd2c30895e005829",
+    "mattermost-redux": "github:mattermost/mattermost-redux#21a808ab7f8deaf5dcac31fc6ca73ed2c7fe2a41",
     "moment-timezone": "0.5.21",
     "pdfjs-dist": "2.0.489",
     "perfect-scrollbar": "0.8.1",
