+++ sass/components/_admin-panel.scss
@@ -44,12 +44,6 @@
         }
     }
 
-    .footer {
-        background: $white;
-        padding: 10px;
-        height: auto;
-    }
-
     &.closed {
         .header {
             .button {
